requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.3
pyyaml>=6.0.1
asyncpg>=0.29.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
tiktoken>=0.6.0
# File Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
